# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateDutyRequest:
    properties:
      assigneeId:
        type: string
      date:
        type: string
      floor:
        type: integer
      type:
        enum:
        - DORM
        - NIGHT_STUDY
        type: string
    required:
    - assigneeId
    - date
    - type
    type: object
...
  dto.DutyResponse:
    properties:
      assignee:
        $ref: '#/definitions/dto.UserResponse'
      completed:
        type: boolean
      createdAt:
        type: string
      date:
        type: string
      floor:
        type: integer
      id:
        type: string
      type:
        type: string
    type: object
  dto.GenerateDutyRequest:
    properties:
      assigneeIds:
        items:
          type: string
        minItems: 1
        type: array
      endDate:
        type: string
      floor:
        type: integer
      startDate:
        type: string
      type:
        enum:
        - DORM
        - NIGHT_STUDY
        type: string
    required:
    - assigneeIds
    - endDate
    - startDate
    - type
    type: object
...
  dto.PaginatedResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/dto.Pagination'
      success:
        type: boolean
    type: object
  dto.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dto.SwapDutyRequest:
    properties:
      targetDutyId:
        type: string
    required:
    - targetDutyId
    type: object
  dto.UpdateDutyRequest:
    properties:
      assigneeId:
        type: string
      date:
        type: string
      floor:
        type: integer
      type:
        enum:
        - DORM
        - NIGHT_STUDY
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 기숙사 관리 시스템 API - 학생 상벌점 관리 및 당직 일정 운영
  title: Dormi API
  version: "1.0"
paths:
...
  /duties:
    get:
      description: 당직 목록 조회 (필터링 지원)
      parameters:
      - description: 유형 (DORM, NIGHT_STUDY)
        in: query
        name: type
        type: string
      - description: 담당자 ID
        in: query
        name: assigneeId
        type: string
      - description: 시작일 (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: 종료일 (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - default: 1
        description: 페이지
        in: query
        name: page
        type: integer
      - default: 20
        description: 페이지당 개수
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.DutyResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 당직 목록 조회
      tags:
      - 당직
    post:
      consumes:
      - application/json
      description: 새로운 당직 일정 생성
      parameters:
      - description: 당직 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDutyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DutyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 당직 생성
      tags:
      - 당직
  /duties/{id}:
    delete:
      description: 당직 정보 삭제
      parameters:
      - description: 당직 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 당직 삭제
      tags:
      - 당직
    get:
      description: ID로 당직 정보 조회
      parameters:
      - description: 당직 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DutyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 당직 상세 조회
      tags:
      - 당직
    put:
      consumes:
      - application/json
      description: 당직 정보 수정
      parameters:
      - description: 당직 ID
        in: path
        name: id
        required: true
        type: string
      - description: 수정할 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDutyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DutyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 당직 수정
      tags:
      - 당직
  /duties/{id}/complete:
    patch:
      description: 당직을 완료 상태로 변경
      parameters:
      - description: 당직 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 당직 완료 처리
      tags:
      - 당직
  /duties/{id}/swap:
    post:
      consumes:
      - application/json
      description: 두 당직의 담당자 교대
      parameters:
      - description: 당직 ID
        in: path
        name: id
        required: true
        type: string
      - description: 교대 대상 당직 ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SwapDutyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 당직 교대
      tags:
      - 당직
  /duties/generate:
    post:
      consumes:
      - application/json
      description: 기간과 담당자 목록으로 당직 자동 생성
      parameters:
      - description: 자동 생성 정보
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GenerateDutyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.DutyResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 당직 자동 생성
      tags:
      - 당직
...
securityDefinitions:
  BearerAuth:
    description: JWT 토큰 (Bearer {token})
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"